# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DaliJob API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/operations/job-search": {
      "post": {
        "tags": [
          "operations"
        ],
        "summary": "Enqueue Job Search",
        "operationId": "enqueue_job_search_api_v1_operations_job_search_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IndeedJobSearchRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagedOperationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/operations/quick-find-jobs": {
      "post": {
        "tags": [
          "operations"
        ],
        "summary": "Enqueue Quick Find Jobs",
        "operationId": "enqueue_quick_find_jobs_api_v1_operations_quick_find_jobs_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuickFindRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagedOperationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/operations/ask-scout": {
      "post": {
        "tags": [
          "operations"
        ],
        "summary": "Enqueue Ask Scout",
        "operationId": "enqueue_ask_scout_api_v1_operations_ask_scout_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskScoutRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagedOperationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/operations/provider-job-import": {
      "post": {
        "tags": [
          "operations"
        ],
        "summary": "Enqueue Provider Job Import",
        "operationId": "enqueue_provider_job_import_api_v1_operations_provider_job_import_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IndeedJobSearchImportRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagedOperationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/operations/job-list-discover": {
      "post": {
        "tags": [
          "operations"
        ],
        "summary": "Enqueue Job List Discover",
        "operationId": "enqueue_job_list_discover_api_v1_operations_job_list_discover_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobListDiscoverRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagedOperationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/operations/job-list-import": {
      "post": {
        "tags": [
          "operations"
        ],
        "summary": "Enqueue Job List Import",
        "operationId": "enqueue_job_list_import_api_v1_operations_job_list_import_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobListImportRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagedOperationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IndeedJobSearchRequest": {
        "properties": {
          "keyword": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Keyword"
          },
          "location": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Location"
          },
          "max_results": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Max Results",
            "default": 10
          },
          "search_criterion_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Search Criterion Id"
          }
        },
        "type": "object",
        "required": [
          "keyword",
          "location"
        ],
        "title": "IndeedJobSearchRequest"
      },
...
      "QuickFindRequest": {
        "properties": {
          "resume_profile_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Resume Profile Id"
          },
          "search_criterion_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Search Criterion Id"
          },
          "location": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Location"
          },
          "keyword": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Keyword"
          },
          "max_results": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Max Results",
            "default": 5
          }
        },
        "type": "object",
        "required": [
          "resume_profile_id",
          "location"
        ],
        "title": "QuickFindRequest"
      },
...
    }
  }
}